Add schema org for rich result
