+++ layouts/incubation.vue
@@ -5,19 +5,3 @@
     <IncubationFooter />
   </div>
 </template>
-
-<script setup lang="ts">
-const route = useRoute();
-const { locale } = useI18n();
-import { meta } from "../content/meta";
-const seoUrl = `${meta.url}${route.fullPath}`;
-useHead(() => ({
-  htmlAttrs: { lang: locale.value },
-  link: [
-    {
-      rel: "canonical",
-      href: seoUrl,
-    },
-  ],
-}));
-</script>
