+++ package.json
@@ -22,6 +22,7 @@
     "graphql": "^16.6.0",
     "graphql-tag": "^2.12.6",
     "nuxt": "^3.3.1",
+    "nuxt-schema-org": "^2.1.3",
     "nuxt-simple-sitemap": "^2.4.22",
     "vue3-carousel": "^0.2.9",
     "vue3-parallax": "^0.1.4"
